# Conflicts:
#	src/tiri/tiri_thread.cpp

+++ src/tiri/jit/src/debug/dump_bytecode.cpp
@@ -303,7 +303,7 @@ void format_bc_line(lua_State *L, BCLine Line, int FileWidth, BytecodeLogger Log
 
          std::string_view sv("<unknown>");
          if (src) sv = std::string_view(src->filename);
-         if (sv.size() > FileWidth) sv.remove_suffix(sv.size() - FileWidth);
+         if (sv.size() > size_t(FileWidth)) sv.remove_suffix(sv.size() - FileWidth);
          file_and_line = std::format("{}:{}", sv, Line.lineNumber());
 
          if (src->total_lines > 10000) line_no_width += 2;
